pyproject.toml
[project]
name = "dev-tools"
version = "0.0.0"
description = "Development tools"
requires-python = ">=3.10"

[dependency-groups]
lint = [
    "isort==6.0.1",
    "pydantic-settings==2.8.1",
    "ruff==0.11.6",
    "mypy==1.15.0",
]

[tool.isort]
balanced_wrapping = true
profile = "black"
force_sort_within_sections = true
force_alphabetical_sort_within_sections = true
line_length = 88
lines_between_types = 1
src_paths = ["services"]
known_first_party = [
    "api",
    "core",
    "db",
    "functional",
    "models",
    "openapi",
    "schemas",
    "services",
    "settings",
    "testdata",
    "utils",
    ]
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
    ]

[tool.mypy]
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
strict = true

[tool.ruff]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
]
line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
